requirements.txt
aiocache==0.11.1
aiocqhttp==1.4.3
aiofile==3.8.1
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
bbcode==1.1.0
binaryornot==0.4.4
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
build==0.9.0
CacheControl==0.12.11
cachy==0.3.0
caio==0.9.8
cashews==5.0.0
cchardet==2.1.7
certifi==2022.9.24
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.7
cookiecutter==1.7.3
crashtest==0.3.1
cycler==0.11.0
distlib==0.3.6
Django==4.1.6
docutils==0.19
dulwich==0.20.50
et-xmlfile==1.1.0
fastapi==0.88.0
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.1
greenlet==1.1.3
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
html5lib==1.1
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
hypercorn==0.14.3
hyperframe==5.2.0
identify==2.5.8
idna==2.10
imgkit==1.2.2
importlib-metadata==4.13.0
install==1.3.5
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.0
keyring==23.9.3
kiwisolver==1.4.4
lockfile==0.12.2
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.3
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.4
mypy==0.982
mypy-extensions==0.4.3
mysqlclient==2.1.1
Naked==0.1.31
nb-cli==1.0.2
nodeenv==1.7.0
nonebot-adapter-onebot==2.2.0
nonebot-plugin-imageutils==0.1.13.5
nonebot2==2.0.0rc3
noneprompt==0.1.7
numpy==1.24.1
opencv-python-headless==4.7.0.68
openpyxl==3.0.10
packaging==21.3
pandas==1.5.3
pep517==0.13.0
pexpect==4.8.0
Pillow==9.4.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.1.2
poyo==0.5.0
pre-commit==2.20.0
prettytable==3.4.1
priority==2.0.0
prompt-toolkit==3.0.31
psutil==5.9.4
ptyprocess==0.7.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.4
pyecharts==1.9.1
pyfiglet==0.8.post1
Pygments==2.13.0
pygtrans==1.4.0
pygtrie==2.5.0
pylev==1.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
quart==0.18.3
readme-renderer==37.3
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.6.0
shellescape==3.8.1
shellingham==1.5.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.22.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
twine==4.0.1
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.20.0
virtualenv==20.17.1
watchfiles==0.17.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.8.1
zipp==3.10.0
